% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{as_tbl_grid}
\alias{as_tbl_grid}
\title{Convert a data frame into a tbl_grid object}
\usage{
as_tbl_grid(x, var = NULL, grid_size = NULL, strict = TRUE, ...)
}
\arguments{
\item{x}{An object to be converted into an object class \code{tbl_grid}.}

\item{var}{A variable to specify the grid object. By default, the first
column of the grid object is taken.}

\item{grid_size}{A grid size.}

\item{strict}{A logical scalar. Should the number of digits in the grid
square code match a given number of digits?}

\item{...}{Additional arguments passed to \code{\link[stickyr:new_sticky_tibble]{stickyr::new_sticky_tibble()}}}
}
\value{
A \code{tbl_grid} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
It is recommended to use \code{grid_as_sf()}.

The \code{tbl_grid} object is a data frame with \code{grid} objects in the columns.
\code{as_tbl_grid} converts a data frame into a tbl_grid object.
}
