% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{bbox_to_mesh}
\alias{bbox_to_mesh}
\title{Converting bbox to regional meshes}
\usage{
bbox_to_mesh(bbox, size)
}
\arguments{
\item{bbox}{A \code{bbox} or a list of \code{bbox}.}

\item{size}{A mesh size.}
}
\value{
A \code{mesh} vector (when bbox is a \code{bbox}) or A list of \code{mesh} vectors (when bbox is a list of \code{bbox}).
}
\description{
Converting bbox to regional meshes
}
