% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_grid}
\alias{convert_grid}
\title{Convert the grid size of grid objects}
\usage{
convert_grid(grid, grid_size)
}
\arguments{
\item{grid}{A \code{grid} vector.}

\item{grid_size}{A grid size.}
}
\value{
A \code{grid} vector.
}
\description{
Convert the grid size of grid objects
}
\examples{
grid_500m <- parse_grid(c("533945263", "533935863", "533945764"), "500m")
convert_grid(grid_500m, "10km")

}
