% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{coords}
\alias{coords}
\alias{coords_to_grid}
\alias{grid_to_coords}
\title{Conversion between grid square codes and coordinates (longitude and latitude)}
\usage{
coords_to_grid(X, Y, grid_size)

grid_to_coords(grid, center = TRUE)
}
\arguments{
\item{X}{A numeric vector of longitude.}

\item{Y}{A numeric vector of latitude.}

\item{grid_size}{A grid size.}

\item{grid}{A \code{grid} class vector.}

\item{center}{Should the center point of the grid be returned? Otherwise the
end points will be returned. \code{TRUE} by default.}
}
\value{
\code{coords_to_grid()} returns a \code{grid} vector.

\code{grid_to_coords()} returns a \code{tbl_df}.
}
\description{
Conversion between grid square codes and coordinates (longitude and latitude)
}
