% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{geometry_to_mesh}
\alias{geometry_to_mesh}
\title{Converting sfc geometries to regional meshes}
\usage{
geometry_to_mesh(geometry, size, ...)
}
\arguments{
\item{geometry}{A \code{sfc} vector.}

\item{size}{A mesh size.}

\item{...}{Passed on to \code{stars::st_rasterize()}.}
}
\value{
A list of \code{mesh} class vectors.
}
\description{
Converting sfc geometries to regional meshes
}
