% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{grid_as_stars}
\alias{grid_as_stars}
\title{Converting data frame containing regional grids to stars}
\usage{
grid_as_stars(
  x,
  coords = NULL,
  crs = sf::NA_crs_,
  grid_column_name = NULL,
  ...
)
}
\arguments{
\item{x}{A data frame.}

\item{grid_column_name}{A scalar character.}

\item{...}{Passed on to \code{stars::st_as_stars()}.}
}
\value{
A \code{stars} object.
}
\description{
Converting data frame containing regional grids to stars
}
