% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{grid_bbox}
\alias{grid_bbox}
\title{Convert grid square codes into bounding codes}
\usage{
grid_bbox(grid)
}
\arguments{
\item{grid}{A \code{grid} vector.}
}
\value{
A \code{grid} vector.
}
\description{
Convert grid square codes into bounding codes
}
