% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{grid_boundary}
\alias{grid_boundary}
\title{Grid square codes on the boundary}
\usage{
grid_boundary(grid, n = 1, type = NULL)
}
\arguments{
\item{grid}{A \code{grid} vector.}

\item{n}{A numeric vector of degrees. By default, \code{1}.}

\item{type}{A character vector of neighborhood types, \code{"von_neumann"} or
\code{"moore"}. By default, \code{"von_neumann"}.}
}
\value{
A \code{grid} vector.
}
\description{
Grid square codes on the boundary
}
