% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{grid_circle}
\alias{grid_circle}
\title{Circular grid square codes}
\usage{
grid_circle(X, Y, dist, grid_size, crs = 4326, ...)
}
\arguments{
\item{X}{A numeric vector of longitude.}

\item{Y}{A numeric vector of latitude.}

\item{dist}{A numeric vector of distances passed on to \code{\link[sf:geos_unary]{sf::st_buffer()}}.}

\item{grid_size}{A grid size.}

\item{crs}{Coordinate reference system.}

\item{...}{Passed on to \code{geometry_to_grid()}.}
}
\value{
A list of \code{grid} vector.
}
\description{
Circular grid square codes
}
