% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_city2015}
\alias{grid_city2015}
\title{List of grid square codes by Japanese municipalities in 2015}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 461373 rows and 4 columns.
}
\source{
\url{http://www.stat.go.jp/data/mesh/m_itiran.html}
}
\usage{
grid_city2015
}
\description{
List of grid square codes by Japanese municipalities in 2015
}
\keyword{datasets}
