% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_city_2020}
\alias{grid_city_2020}
\title{List of grid square codes by Japanese municipalities}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 462915 rows and 6 columns.
}
\source{
\url{https://www.stat.go.jp/data/mesh/m_itiran.html}
}
\usage{
grid_city_2020
}
\description{
List of grid square codes by Japanese municipalities
}
\keyword{datasets}
