% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{grid_from_bbox}
\alias{grid_from_bbox}
\title{Converting bbox to grid square codes}
\usage{
grid_from_bbox(bbox, grid_size)
}
\arguments{
\item{bbox}{A \code{bbox}.}

\item{grid_size}{A grid size.}
}
\value{
A \code{grid} vector.
}
\description{
Converting bbox to grid square codes
}
