% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{grid_from_geometry}
\alias{grid_from_geometry}
\title{Converting sfc geometries to grid square codes}
\usage{
grid_from_geometry(geometry, size, options = "ALL_TOUCHED=TRUE", ...)
}
\arguments{
\item{geometry}{A \code{sfc} vector.}

\item{size}{A grid size.}

\item{options}{Options vector for GDALRasterize passed on to
\code{\link[stars:st_rasterize]{stars::st_rasterize()}}.}

\item{...}{Passed on to \code{\link[stars:st_rasterize]{stars::st_rasterize()}}.}
}
\value{
A list of \code{grid} vectors.
}
\description{
Converting sfc geometries to grid square codes
}
