% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{grid_neighbor}
\alias{grid_neighbor}
\title{Neighborhood grid square codes}
\usage{
grid_neighbor(grid, n = 1L, moore = TRUE, simplify = TRUE)
}
\arguments{
\item{grid}{A \code{grid} vector.}

\item{n}{A numeric vector of degrees.}

\item{moore}{Moore neighborhood (\code{TRUE}) or Von Neumann neighborhood
(\code{FALSE}).}

\item{simplify}{Should simplify the format of the return?}
}
\value{
A list of \code{grid} vectors.
}
\description{
Neighborhood grid square codes
}
