% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{grid_neighborhood}
\alias{grid_neighborhood}
\title{Neighborhood grid square codes}
\usage{
grid_neighborhood(grid, n = 1L, type = NULL, simplify = TRUE)
}
\arguments{
\item{grid}{A \code{grid} vector.}

\item{n}{A numeric vector of degrees. By default, \code{1L}.}

\item{type}{A character vector of neighborhood types, \code{"von_neumann"} or
\code{"moore"}. By default, \code{"von_neumann"}.}

\item{simplify}{Should simplify the format of the return?}
}
\value{
A list of \code{grid} vectors.
}
\description{
Neighborhood grid square codes
}
