% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{grid_parse}
\alias{grid_parse}
\title{Parse grid square codes}
\usage{
grid_parse(x, size = NULL, strict = TRUE)
}
\arguments{
\item{x}{A character vector of grid square codes.}

\item{size}{A grid size.}

\item{strict}{A logical scalar. Should the number of digits in the grid
square code match a given number of digits? By default, \code{TRUE}.}
}
\description{
Parse grid square codes
}
\examples{
grid_parse("53394526313")
grid_parse("53394526313",
           size = "80km")
grid_parse("53394526313",
           size = "80km",
           strict = FALSE)

}
