% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangle.R
\name{grid_rectangle}
\alias{grid_rectangle}
\title{Convert regional grids into rectangular grids}
\usage{
grid_rectangle(grid)
}
\arguments{
\item{grid}{A \code{grid} vector.}
}
\value{
A \code{grid} vector.
}
\description{
Convert regional grids into rectangular grids
}
