% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdivide.R
\name{grid_subdivide}
\alias{grid_subdivide}
\title{Subdivide regional grides}
\usage{
grid_subdivide(grid, size)
}
\arguments{
\item{grid}{A \code{grid} vector.}

\item{size}{A grid size.}
}
\value{
A list of \code{grid} class vector.
}
\description{
\code{grid_subdivide} makes the regional grides finer.
}
