% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{is_grid}
\alias{is_grid}
\title{Test if the object is a grid}
\usage{
is_grid(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{grid} class.
}
\description{
Test if the object is a grid
}
