% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpgrid.R
\docType{package}
\name{japangrid}
\alias{japanmesh}
\alias{japanmesh-package}
\title{Functions for the Japanese Regional grid Codes (JIS X 0410)}
\description{
japangrid is an R package for using the reference regional grid (the 1st grid to the 3rd grid),
the split regional grid as defined by the JIS (Japan Industrial Standard) X 0410 'regional grid code'
and 1/10 subdivision of the 3rd grid.
Regional grid codes are square-like regional divisions set up for all regions of Japan based on longitude and latitude.
}
\seealso{
\url{https://www.jisc.go.jp/app/jis/general/GnrJISNumberNameSearchList?show&jisStdNo=X0410}

\url{https://www.stat.go.jp/data/grid/pdf/gaiyo1.pdf}
}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
