% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/japanmesh.R
\docType{package}
\name{japanmesh}
\alias{japanmesh}
\alias{japanmesh-package}
\title{Functions for the Japanese Regional Mesh Codes (JIS X 0410)}
\description{
japanmesh is an R package for using the reference regional mesh (the 1st mesh to the 3rd mesh),
the split regional mesh as defined by the JIS (Japan Industrial Standard) X 0410 'regional mesh code'
and 1/10 subdivision of the 3rd mesh.
Regional mesh codes are square-like regional divisions set up for all regions of Japan based on longitude and latitude.
}
\seealso{
\url{https://www.jisc.go.jp/app/jis/general/GnrJISNumberNameSearchList?show&jisStdNo=X0410}

\url{https://www.stat.go.jp/data/mesh/pdf/gaiyo1.pdf}
}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
