% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpgrid-package.R
\docType{package}
\name{jpgrid}
\alias{jpgrid-package}
\alias{jpgrid}
\title{Functions for the Grid Square Codes in Japan}
\description{
Provides functions for grid square codes in Japan (\url{https://www.stat.go.jp/english/data/mesh/index.html}). Generates the grid square codes from longitude/latitude, geometries, and the grid square codes of different scales, and vice versa.
}
\seealso{
\url{https://www.stat.go.jp/english/data/mesh/index.html}
}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
