% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{mesh}
\alias{mesh}
\alias{as_mesh}
\alias{as_mesh.default}
\alias{as_mesh.mesh}
\alias{mesh_80km}
\alias{mesh_10km}
\alias{mesh_1km}
\alias{mesh_500m}
\alias{mesh_250m}
\alias{mesh_125m}
\alias{mesh_100m}
\alias{is_mesh}
\alias{format.mesh_80km}
\alias{format.mesh_10km}
\alias{format.mesh_1km}
\alias{format.mesh_500m}
\alias{format.mesh_250m}
\alias{format.mesh_125m}
\alias{format.mesh_100m}
\title{Meshcode vector}
\usage{
as_mesh(x, size, ...)

\method{as_mesh}{default}(x, size, ...)

\method{as_mesh}{mesh}(x, size = NULL, ...)

mesh_80km(x)

mesh_10km(x)

mesh_1km(x)

mesh_500m(x)

mesh_250m(x)

mesh_125m(x)

mesh_100m(x)

is_mesh(x)

\method{format}{mesh_80km}(x, ...)

\method{format}{mesh_10km}(x, ...)

\method{format}{mesh_1km}(x, ...)

\method{format}{mesh_500m}(x, ...)

\method{format}{mesh_250m}(x, ...)

\method{format}{mesh_125m}(x, ...)

\method{format}{mesh_100m}(x, ...)
}
\arguments{
\item{x}{meshcode}

\item{size}{meshcode size}
}
\description{
Meshcode vector
}
\examples{
mesh_80km("53394526313")
mesh_1km(c("53394526313", "5339358633", "533945764"))

}
