% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{mesh_as_sf}
\alias{mesh_as_sf}
\title{Converting data frame containing regional meshes to sf}
\usage{
mesh_as_sf(
  x,
  as_points = FALSE,
  crs = sf::NA_crs_,
  mesh_column_name = NULL,
  ...
)
}
\arguments{
\item{x}{A data frame.}

\item{as_points}{Return the center points of the meshes or not?}

\item{crs}{Coordinate reference system.}

\item{mesh_column_name}{A scalar character.}

\item{...}{passed on to \code{sf::st_as_sf()}.}
}
\value{
A \code{sf} object.
}
\description{
Converting data frame containing regional meshes to sf
}
