% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{mesh_as_sfc}
\alias{mesh_as_sfc}
\title{Converting regional meshes to sfc geometries}
\usage{
mesh_as_sfc(x, as_points = FALSE, crs = sf::NA_crs_)
}
\arguments{
\item{x}{A \code{mesh} vector.}

\item{as_points}{Return the center points of the meshes or not?}

\item{crs}{Coordinate reference system.}
}
\value{
A \code{sfc_POLYGON} vector (\code{as_points = FALSE}) or a \code{sfc_POINT} vector ((\code{as_points = TRUE})).
}
\description{
Converting regional meshes to sfc geometries
}
