% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{mesh_as_stars}
\alias{mesh_as_stars}
\title{Converting data frame containing regional meshes to stars}
\usage{
mesh_as_stars(x, mesh_column_name = NULL, ...)
}
\arguments{
\item{x}{A data frame.}

\item{mesh_column_name}{A scalar character.}

\item{...}{passed on to \code{stars::st_as_stars()}.}
}
\value{
A \code{stars} object.
}
\description{
Converting data frame containing regional meshes to stars
}
