% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{mesh_move}
\alias{mesh_move}
\title{Moving on the mesh}
\usage{
mesh_move(mesh, n_X, n_Y)
}
\arguments{
\item{mesh}{A \code{mesh} vector.}

\item{n_X}{Number of moving cells in the longitude direction.}

\item{n_Y}{Number of moving cells in the latitude direction.}
}
\description{
Moving on the mesh
}
