% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{mesh_neighbor}
\alias{mesh_neighbor}
\title{Neighbor mesh}
\usage{
mesh_neighbor(mesh, n = 1L, moore = TRUE, simplify = TRUE)
}
\arguments{
\item{mesh}{A \code{mesh} vector.}

\item{n}{A numeric vector of degrees.}

\item{moore}{Moore neighborhood (\code{TRUE}) or Von Neumann neighborhood (\code{FALSE}).}

\item{simplify}{Should simplify the format of the return?}
}
\description{
Neighbor mesh
}
