% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangle.R
\name{mesh_rectangle}
\alias{mesh_rectangle}
\title{Convert regional meshes into rectangular meshes}
\usage{
mesh_rectangle(mesh, buffer = 0L)
}
\arguments{
\item{mesh}{A \code{mesh} vector.}
}
\value{
A \code{mesh} vector.
}
\description{
Convert regional meshes into rectangular meshes
}
