% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdivide.R
\name{mesh_subdivide}
\alias{mesh_subdivide}
\title{Subdivide regional meshes}
\usage{
mesh_subdivide(mesh, size)
}
\arguments{
\item{mesh}{A \code{mesh} vector.}

\item{size}{A mesh size.}
}
\value{
A list of \code{mesh} class vector.
}
\description{
\code{mesh_subdivide} makes the regional meshes finer.
}
