% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{mesh_to_geometry}
\alias{mesh_to_geometry}
\alias{mesh_to_polygon}
\alias{mesh_to_point}
\title{Converting regional meshes to sfc geometries}
\usage{
mesh_to_polygon(mesh, crs = sf::NA_crs_)

mesh_to_point(mesh, crs = sf::NA_crs_)
}
\arguments{
\item{mesh}{A \code{mesh} vector.}

\item{crs}{Coordinate reference system.}
}
\value{
\code{mesh_to_polygon} returns a \code{sfc_POLYGON} vector.
\code{mesh_to_point} returns a \code{sfc_POINT} vector.
}
\description{
Converting regional meshes to sfc geometries
}
