% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{mesh_to_polygon}
\alias{mesh_to_polygon}
\title{Converting regional meshes to sfc polygons}
\usage{
mesh_to_polygon(mesh, crs = sf::NA_crs_)
}
\arguments{
\item{mesh}{A \code{mesh} vector.}

\item{crs}{Coordinate reference system.}
}
\value{
A \code{sfc_POLYGON} vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use \code{mesh_as_sfc()} instead of \code{mesh_to_polygon()}.
}
