% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{mesh_zoomin}
\alias{mesh_zoomin}
\title{Zoom-in regional meshes}
\usage{
mesh_zoomin(mesh, size)
}
\arguments{
\item{mesh}{A \code{mesh} vector.}

\item{size}{A mesh size.}
}
\value{
A list of \code{mesh} class vector.
}
\description{
`r lifecycle::badge("deprecated")`
\code{mesh_zoomin} makes the regional meshes finer.
}
