% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{mesh_zoomout}
\alias{mesh_zoomout}
\title{Zoom-out regional meshes}
\usage{
mesh_zoomout(mesh, size)
}
\arguments{
\item{mesh}{A \code{mesh} vector.}

\item{size}{A mesh size.}
}
\value{
A \code{mesh} class vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{mesh_zoomout} makes the regional meshes coarser.
}
