% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move_grid}
\alias{move_grid}
\title{Moving on grid square codes}
\usage{
move_grid(grid, n_X, n_Y)
}
\arguments{
\item{grid}{A \code{grid} vector.}

\item{n_X}{Number of moving cells in the longitude direction.}

\item{n_Y}{Number of moving cells in the latitude direction.}
}
\value{
A \code{grid} vector.
}
\description{
Moving on grid square codes
}
