% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_grid}
\alias{parse_grid}
\title{Parse grid square codes}
\usage{
parse_grid(x, grid_size = NULL, strict = TRUE)
}
\arguments{
\item{x}{A character vector of grid square codes.}

\item{grid_size}{A grid size.}

\item{strict}{A scalar logical. Should the number of digits in the grid
square code match a given number of digits? By default, \code{TRUE}.}
}
\description{
Parse grid square codes
}
\examples{
parse_grid("53394526313")
parse_grid("53394526313", "80km")
parse_grid("53394526313", "80km",
           strict = FALSE)

}
