% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{point_to_grid}
\alias{point_to_grid}
\title{Converting sfc points to regional grids}
\usage{
point_to_grid(point, size)
}
\arguments{
\item{point}{A \code{sfc_POINT} vector.}

\item{size}{A grid size.}
}
\value{
A \code{grid} class vector.
}
\description{
Converting sfc points to regional grids
}
