% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sfc_to_mesh}
\alias{sfc_to_mesh}
\title{Converting sfc polygons to regional meshes}
\usage{
sfc_to_mesh(sfc, size, .predicate = sf::st_intersects)
}
\arguments{
\item{sfc}{A \code{sfc} vector.}

\item{size}{A mesh size.}

\item{.predicate}{A \code{.predicate} parameter for \code{sf::st_filter} function.}
}
\value{
A list of \code{mesh} class vectors.
}
\description{
Converting sfc polygons to regional meshes
}
