% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XY.R
\name{XY}
\alias{XY}
\alias{XY_to_mesh}
\alias{mesh_to_XY}
\title{Conversion between mesh and coordinates (longitude and latitude)}
\usage{
XY_to_mesh(X, Y, size)

mesh_to_XY(mesh, center = T, ...)
}
\arguments{
\item{X}{A numeric vector of longitude.}

\item{Y}{A numeric vector of latitude.}

\item{size}{A mesh size.}

\item{mesh}{A \code{mesh} class vector.}

\item{center}{Should the center point of the mesh be returned? Otherwise the end points will be returned.}
}
\value{
\code{XY_to_mesh} returns a \code{mesh} class vector.

\code{mesh_to_XY} returns a tbl.
}
\description{
Conversion between mesh and coordinates (longitude and latitude)
}
