% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{XY_to_grid}
\alias{XY_to_grid}
\title{Conversion between grid square codes and coordinates (longitude and latitude)}
\usage{
XY_to_grid(X, Y, size)
}
\arguments{
\item{X}{A numeric vector of longitude.}

\item{Y}{A numeric vector of latitude.}

\item{size}{A grid size.}
}
\value{
\code{XY_to_grid} returns a \code{grid} vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
