% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoom}
\alias{zoom}
\alias{mesh_zoomout}
\alias{mesh_zoomin}
\title{Zoom meshcode}
\usage{
mesh_zoomout(mesh, size)

mesh_zoomin(mesh, size)
}
\arguments{
\item{mesh}{meshcode}

\item{size}{meshcode size}
}
\value{
meshcode

list of meshcode
}
\description{
Zoom meshcode
}
