% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{bbox_to_grid}
\alias{bbox_to_grid}
\title{Converting bbox to grid square codes}
\usage{
bbox_to_grid(bbox, grid_size)
}
\arguments{
\item{bbox}{A \code{bbox}.}

\item{grid_size}{A grid size.}
}
\value{
A \code{grid} vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
It is recommended to use \code{grid_from_bbox()}.
}
