% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{bbox_to_grid}
\alias{bbox_to_grid}
\title{Converting bbox to grid square codes}
\usage{
bbox_to_grid(bbox, size)
}
\arguments{
\item{bbox}{A \code{bbox}.}

\item{size}{A grid size.}
}
\value{
A \code{grid} vector.
}
\description{
Converting bbox to grid square codes
}
