% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{geometry_to_grid}
\alias{geometry_to_grid}
\title{Converting sfc geometries to grid square codes}
\usage{
geometry_to_grid(geometry, size, options = "ALL_TOUCHED=TRUE", ...)
}
\arguments{
\item{geometry}{A \code{sfc} vector.}

\item{size}{A grid size.}

\item{options}{Options vector for GDALRasterize passed on to
\code{\link[stars:st_rasterize]{stars::st_rasterize()}}.}

\item{...}{Passed on to \code{\link[stars:st_rasterize]{stars::st_rasterize()}}.}
}
\value{
A list of \code{grid} vectors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
It is recommended to use \code{grid_from_geom()}.
}
