% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{geometry_to_grid}
\alias{geometry_to_grid}
\title{Converting sfc geometries to grid square codes}
\usage{
geometry_to_grid(geometry, size, ...)
}
\arguments{
\item{geometry}{A \code{sfc} vector.}

\item{size}{A grid size.}

\item{...}{Passed on to \code{\link[stars:st_rasterize]{stars::st_rasterize()}}.}
}
\value{
A list of \code{grid} vectors.
}
\description{
Converting sfc geometries to grid square codes
}
