% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{geometry_to_mesh}
\alias{geometry_to_mesh}
\title{Converting geometries to regional meshes}
\usage{
geometry_to_mesh(geometry, size, .predicate = sf::st_intersects)
}
\arguments{
\item{geometry}{A \code{sfc} vector.}

\item{size}{A mesh size.}

\item{.predicate}{A \code{.predicate} parameter for \code{sf::st_filter} function.}
}
\value{
A list of \code{mesh} class vectors.
}
\description{
Converting geometries to regional meshes
}
