% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{grid_class}
\alias{grid_class}
\alias{grid_80km}
\alias{grid_10km}
\alias{grid_1km}
\alias{grid_500m}
\alias{grid_250m}
\alias{grid_125m}
\alias{grid_100m}
\alias{grid_auto}
\title{Grid square code vector}
\usage{
grid_80km(x, strict = TRUE)

grid_10km(x, strict = TRUE)

grid_1km(x, strict = TRUE)

grid_500m(x, strict = TRUE)

grid_250m(x, strict = TRUE)

grid_125m(x, strict = TRUE)

grid_100m(x, strict = TRUE)

grid_auto(x, strict = TRUE)
}
\arguments{
\item{x}{A list or vector.}

\item{strict}{A logical scalar. Should the number of digits in the grid
square code match a given number of digits?}
}
\value{
A \code{grid} vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
It is recommended to use \code{grid_parse()} or \code{grid_convert()}.

A series of functions return \code{grid} class for each grid size.
\code{grid_auto()} returns automatically determine grid size by the largest
grid size.
}
