% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{grid_components}
\alias{grid_components}
\title{Connected components of grid square codes}
\usage{
grid_components(grid, n = 0:1, moore = FALSE)
}
\arguments{
\item{grid}{A \code{grid} vector.}

\item{n}{A numeric vector of degrees. By default, \code{0:1}.}

\item{moore}{Moore neighborhood (\code{TRUE}) or Von Neumann neighborhood
(\code{FALSE}, default).}
}
\value{
A integer vector of group IDs.
}
\description{
Connected components of grid square codes
}
