% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{grid_convert}
\alias{grid_convert}
\title{Convert the grid size of grid objects}
\usage{
grid_convert(grid, size)
}
\arguments{
\item{grid}{A \code{grid} vector.}

\item{size}{A grid size.}
}
\value{
A \code{grid} vector.
}
\description{
Convert the grid size of grid objects
}
\examples{
grid_500m <- grid_parse(c("533945263", "533935863", "533945764"),
                        size = "500m")
grid_convert(grid_500m,
             size = "10km")

}
