% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{grid_neighbor}
\alias{grid_neighbor}
\title{Neighborhood grid square codes (Deprecated)}
\usage{
grid_neighbor(grid, n = 1L, moore = TRUE, simplify = TRUE)
}
\arguments{
\item{grid}{A \code{grid} vector.}

\item{n}{A numeric vector of degrees. By default, \code{1L}.}

\item{moore}{Moore neighborhood (\code{TRUE}, default) or Von Neumann neighborhood
(\code{FALSE}).}

\item{simplify}{Should simplify the format of the return?}
}
\value{
A list of \code{grid} vectors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
