% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{mesh}
\alias{mesh}
\alias{mesh_80km}
\alias{mesh_10km}
\alias{mesh_1km}
\alias{mesh_500m}
\alias{mesh_250m}
\alias{mesh_125m}
\alias{mesh_100m}
\alias{mesh_auto}
\alias{is_mesh}
\alias{format.mesh_80km}
\alias{format.mesh_10km}
\alias{format.mesh_1km}
\alias{format.mesh_500m}
\alias{format.mesh_250m}
\alias{format.mesh_125m}
\alias{format.mesh_100m}
\title{Meshcode vector}
\usage{
mesh_80km(x, strict = T)

mesh_10km(x, strict = T)

mesh_1km(x, strict = T)

mesh_500m(x, strict = T)

mesh_250m(x, strict = T)

mesh_125m(x, strict = T)

mesh_100m(x, strict = T)

mesh_auto(x, strict = T)

is_mesh(x)

\method{format}{mesh_80km}(x, ...)

\method{format}{mesh_10km}(x, ...)

\method{format}{mesh_1km}(x, ...)

\method{format}{mesh_500m}(x, ...)

\method{format}{mesh_250m}(x, ...)

\method{format}{mesh_125m}(x, ...)

\method{format}{mesh_100m}(x, ...)
}
\arguments{
\item{x}{meshcode}
}
\description{
Meshcode vector
}
\examples{
mesh_80km("53394526313")
mesh_1km(c("53394526313", "5339358633", "533945764"))
}
