% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{mesh_grid}
\alias{mesh_grid}
\title{Create regional mesh grids}
\usage{
mesh_grid(X_min, Y_min, X_max, Y_max, size)
}
\arguments{
\item{X_min}{A numeric vector.}

\item{Y_min}{A numeric vector.}

\item{X_max}{A numeric vector.}

\item{Y_max}{A numeric vector.}

\item{size}{A mesh size.}
}
\value{
A list of \code{mesh} vectors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
