% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{mesh_zoomin}
\alias{mesh_zoomin}
\title{Zoom-in regional meshes}
\usage{
mesh_zoomin(mesh, size)
}
\arguments{
\item{mesh}{A \code{mesh} vector.}

\item{size}{A mesh size.}
}
\value{
\code{mesh_zoomin} returns a list of \code{mesh} class vector.
}
\description{
\code{mesh_zoomin} makes the regional meshes finer.
}
