% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{mesh_zoomout}
\alias{mesh_zoomout}
\title{Zoom-out mesh}
\usage{
mesh_zoomout(mesh, size)
}
\arguments{
\item{mesh}{A \code{mesh} vector.}

\item{size}{A mesh size.}
}
\value{
\code{mesh_zoomout} returns a \code{mesh} class vector.
}
\description{
\code{mesh_zoomout} makes the mesh coarser.
}
