% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{point_to_mesh}
\alias{point_to_mesh}
\title{Converting sfc points to meshes}
\usage{
point_to_mesh(point, size)
}
\arguments{
\item{point}{A \code{sfc_POINT} vector.}

\item{size}{A mesh size.}
}
\value{
A \code{mesh} class vector.
}
\description{
Converting sfc points to meshes
}
