% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{XY}
\alias{XY}
\alias{grid_to_XY}
\alias{XY_to_grid}
\title{Conversion between grid square codes and coordinates (longitude and latitude)}
\usage{
grid_to_XY(grid, center = TRUE)

XY_to_grid(X, Y, size)
}
\arguments{
\item{grid}{A \code{grid} class vector.}

\item{center}{Should the center point of the grid be returned? Otherwise the
end points will be returned. \code{TRUE} by default.}

\item{X}{A numeric vector of longitude.}

\item{Y}{A numeric vector of latitude.}

\item{size}{A grid size.}
}
\value{
\code{grid_to_XY()} returns a \code{tbl_df}.

\code{XY_to_grid()} returns a \code{grid} vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
