% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XY.R
\name{XY}
\alias{XY}
\alias{grid_from_XY}
\alias{grid_to_XY}
\title{Conversion between grid square codes and coordinates (longitude and latitude)}
\usage{
grid_from_XY(X, Y, size)

grid_to_XY(grid, center = TRUE)
}
\arguments{
\item{X}{A numeric vector of longitude.}

\item{Y}{A numeric vector of latitude.}

\item{size}{A grid size.}

\item{grid}{A \code{grid} class vector.}

\item{center}{Should the center point of the grid be returned? Otherwise the
end points will be returned. \code{TRUE} by default.}
}
\value{
\code{grid_from_XY} returns a \code{grid} vector.

\code{grid_to_XY} returns a \code{tbl_df}.
}
\description{
Conversion between grid square codes and coordinates (longitude and latitude)
}
